feat: bot now with rag and in docker-compose
