+++ llm_in_cb/bot/config_api.json
