+++ pyproject.toml
@@ -14,6 +14,9 @@ fastapi = "^0.115.5"
 transformers = "^4.46.3"
 uvicorn = "^0.32.1"
 sentence-transformers = "^3.3.1"
+asyncio = "^3.4.3"
+requests = "^2.32.3"
+python-telegram-bot = "^21.7"
 
 
 [build-system]
